requirements.txt
cachetools==4.0.0
certifi==2019.11.28
chardet==3.0.4
cycler==0.10.0
decorator==4.4.1
google-api-core==1.16.0
google-auth==1.10.1
google-cloud-vision==0.41.0
googleapis-common-protos==1.51.0
grpcio==1.26.0
idna==2.8
imageio==2.6.1
kiwisolver==1.1.0
matplotlib==3.1.2
networkx==2.4
numpy==1.18.1
opencv-python==4.1.2.30
Pillow==7.0.0
protobuf==3.11.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.6
python-dateutil==2.8.1
pytz==2019.3
PyWavelets==1.1.1
requests==2.22.0
rsa==4.0
scikit-image==0.16.2
scipy==1.4.1
six==1.14.0
urllib3==1.25.7
